Numpy assignment NumPy Mastery: 10 Theoretical & 10 Numerical Questions
his repository is designed to help learners and practitioners strengthen their understanding of NumPy, one of the most powerful and widely used libraries for numerical computing in Python. It contains a collection of 10 theoretical questions to assess and deepen your conceptual knowledge of NumPy, along with 10 numerical problems to give hands-on experience with real-world data manipulations, array operations, and mathematical computations. Whether you're preparing for interviews, strengthening your Python skills, or deepening your understanding of numerical computing, this repository provides an excellent resource for practical and theoretical learning.